Moved place page i18n messages to their own files (#4921)
