+++ static/js/place/dev_place_overview.tsx
@@ -22,8 +22,9 @@ import React, { useRef } from "react";
 import { LocationCity } from "../components/elements/icons/location_city";
 import { GoogleMap } from "../components/google_map";
 import { formatDate, formatNumber, intl } from "../i18n/i18n";
+import { pageMessages } from "../i18n/i18n_place_messages";
 import { NamedTypedPlace } from "../shared/types";
-import { isPlaceContainedInUsa, pageMessages } from "./util";
+import { isPlaceContainedInUsa } from "./util";
 
 /**
  * Component that displays a table of key demographic statistics for a place.
@@ -83,7 +84,7 @@ const PlaceOverviewTable = (props: {
         <thead>
           <tr>
             <th scope="col" colSpan={2}>
-              Key Demographics
+              {intl.formatMessage(pageMessages.KeyDemographics)}
             </th>
             <th scope="col"></th>
           </tr>
